[New Feature] (test) Add new integration testing of Zookeeper distributed lock features.

+++ test/integration_test/_utils/zookeeper.py
@@ -1,13 +1,18 @@
+import time
+
 from smoothcrawler_cluster._utils.zookeeper import (
     _BaseZookeeperNode,
+    ZookeeperRecipe,
     _BaseZookeeperClient, ZookeeperClient,
     _BaseZookeeperListener
 )
 from kazoo.client import KazooClient
 from kazoo.exceptions import NodeExistsError
 from typing import Type, TypeVar, Generic
+import threading
 import pytest
 import random
+import json
 
 from ..._config import Zookeeper_Hosts
 from ..._values import Test_Zookeeper_Path, Test_Zookeeper_Not_Exist_Path, Test_Zookeeper_String_Value, Test_Zookeeper_Bytes_Value
@@ -30,15 +35,27 @@ class TestZookeeperClient:
         return ZookeeperClient(hosts=Zookeeper_Hosts)
 
     @staticmethod
-    def _remove_path_finally(test_item):
-
+    def _reset_zk_env(test_item):
         def _(self, zk_cli: Generic[_BaseZookeeperClientType]):
             try:
                 test_item(self, zk_cli)
             finally:
-                # Remove the metadata of target path in Zookeeper
-                self.__PyTest_ZK_Client.delete(path=Test_Zookeeper_Path)
+                print(f"[DEBUG _reset_zk_env] self.__PyTest_ZK_Client.exists(path=Test_Zookeeper_Path): {self.__PyTest_ZK_Client.exists(path=Test_Zookeeper_Path)}")
+                if self.__PyTest_ZK_Client.exists(path=Test_Zookeeper_Path) is not None:
+                    # Remove the metadata of target path in Zookeeper
+                    self.__PyTest_ZK_Client.delete(path=Test_Zookeeper_Path)
+        return _
 
+    @staticmethod
+    def _remove_path_finally(test_item):
+        def _(self, zk_cli: Generic[_BaseZookeeperClientType]):
+            try:
+                test_item(self, zk_cli)
+            finally:
+                print(f"[DEBUG _remove_path_finally] self.__PyTest_ZK_Client.exists(path=Test_Zookeeper_Path): {self.__PyTest_ZK_Client.exists(path=Test_Zookeeper_Path)}")
+                if self.__PyTest_ZK_Client.exists(path=Test_Zookeeper_Path) is not None:
+                    # Remove the metadata of target path in Zookeeper
+                    self.__PyTest_ZK_Client.delete(path=Test_Zookeeper_Path, recursive=True)
         return _
 
     @_remove_path_finally
@@ -46,6 +63,13 @@ class TestZookeeperClient:
         # Test with a path which already exists ---> function should return True.
         zk_cli.create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_String_Value)
         assert zk_cli.exist_node(path=Test_Zookeeper_Path) is not None, "It should exist the path (node) it created."
+
+    @_remove_path_finally
+    def test_restrict_exist_path_with_exist_path(self, zk_cli: Generic[_BaseZookeeperClientType]):
+        # Test with a path which already exists ---> function should return True.
+        zk_cli.create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_String_Value)
+        assert zk_cli.restrict_exist_node(path=Test_Zookeeper_Path, restrict=ZookeeperRecipe.ReadLock, identifier="pytest") is not None, \
+            "It should exist the path (node) it created."
 
     def test_exist_path_with_not_exist_path(self, zk_cli: Generic[_BaseZookeeperClientType]):
         # Test with a path which doesn't exist ---> function should return False.
@@ -58,6 +82,16 @@ class TestZookeeperClient:
         assert _creating_result is not None, "It should create a path (Zookeeper node) successfully."
 
         _zk_path = zk_cli.get_node(path=Test_Zookeeper_Path)
+        assert _zk_path.path == Test_Zookeeper_Path, f"The path of *ZookeeperPath* should be equal to {Test_Zookeeper_Path}."
+        assert _zk_path.value == Test_Zookeeper_String_Value, f"The value of *ZookeeperPath* should be equal to \"{Test_Zookeeper_String_Value}\"."
+
+    @_remove_path_finally
+    def test_restrict_get_path_with_exist_path(self, zk_cli: Generic[_BaseZookeeperClientType]):
+        # Test with a path which already exists ---> function should return a object which is _BaseZookeeperClientType type.
+        _creating_result = zk_cli.create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_String_Value)
+        assert _creating_result is not None, "It should create a path (Zookeeper node) successfully."
+
+        _zk_path = zk_cli.restrict_get_node(path=Test_Zookeeper_Path, restrict=ZookeeperRecipe.ReadLock, identifier="pytest")
         assert _zk_path.path == Test_Zookeeper_Path, f"The path of *ZookeeperPath* should be equal to {Test_Zookeeper_Path}."
         assert _zk_path.value == Test_Zookeeper_String_Value, f"The value of *ZookeeperPath* should be equal to \"{Test_Zookeeper_String_Value}\"."
 
@@ -122,6 +156,16 @@ class TestZookeeperClient:
             else:
                 assert False, "It should raise an exception about 'TypeError' because the data type of function argument *value* is invalid."
 
+    # @_reset_zk_env
+    # @_remove_path_finally
+    # def test_restrict_create_path_with_str_value(self, zk_cli: Generic[_BaseZookeeperClientType]):
+    #     _creating_result = zk_cli.restrict_create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_String_Value, restrict=ZookeeperRecipe.WriteLock, identifier="pytest")
+    #     assert _creating_result is not None, "It should create a path (Zookeeper node) successfully."
+    #
+    #     _zk_path = zk_cli.get_node(path=Test_Zookeeper_Path)
+    #     assert _zk_path.path == Test_Zookeeper_Path, f"The path of *ZookeeperPath* should be equal to {Test_Zookeeper_Path}."
+    #     assert _zk_path.value == Test_Zookeeper_String_Value, f"The value of *ZookeeperPath* should be equal to {Test_Zookeeper_String_Value}."
+
     @_remove_path_finally
     def test_get_value_from_path(self, zk_cli: Generic[_BaseZookeeperClientType]):
         _creating_result = zk_cli.create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_String_Value)
@@ -131,12 +175,32 @@ class TestZookeeperClient:
         assert _value == Test_Zookeeper_String_Value, f"The value of *ZookeeperPath* should be equal to {Test_Zookeeper_String_Value}."
 
     @_remove_path_finally
+    def test_restrict_get_value_from_path(self, zk_cli: Generic[_BaseZookeeperClientType]):
+        _creating_result = zk_cli.create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_String_Value)
+        assert _creating_result is not None, "It should create a path (Zookeeper node) successfully."
+
+        _value = zk_cli.restrict_get_value_from_node(path=Test_Zookeeper_Path, restrict=ZookeeperRecipe.ReadLock, identifier="pytest")
+        assert _value == Test_Zookeeper_String_Value, f"The value of *ZookeeperPath* should be equal to {Test_Zookeeper_String_Value}."
+
+    @_remove_path_finally
     def test_set_value_to_path(self, zk_cli: Generic[_BaseZookeeperClientType]):
         _creating_result = zk_cli.create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_String_Value)
         assert _creating_result is not None, "It should create a path (Zookeeper node) successfully."
 
         _new_value = "new zookeeper value"
         zk_cli.set_value_to_node(path=Test_Zookeeper_Path, value=_new_value)
+        assert zk_cli.get_node(path=Test_Zookeeper_Path).value == _new_value, \
+            f"The new value of *ZookeeperPath* should be equal to {_new_value} (by function *get_path*) because it has been modified."
+        assert zk_cli.get_value_from_node(path=Test_Zookeeper_Path) == _new_value, \
+            f"The new value of *ZookeeperPath* should be equal to {_new_value} (by function *get_value_from_path*) because it has been modified."
+
+    @_remove_path_finally
+    def test_restrict_set_value_to_path(self, zk_cli: Generic[_BaseZookeeperClientType]):
+        _creating_result = zk_cli.create_node(path=Test_Zookeeper_Path, value=Test_Zookeeper_String_Value)
+        assert _creating_result is not None, "It should create a path (Zookeeper node) successfully."
+
+        _new_value = "new zookeeper value"
+        zk_cli.restrict_set_value_to_node(path=Test_Zookeeper_Path, value=_new_value, identifier="pytest")
         assert zk_cli.get_node(path=Test_Zookeeper_Path).value == _new_value, \
             f"The new value of *ZookeeperPath* should be equal to {_new_value} (by function *get_path*) because it has been modified."
         assert zk_cli.get_value_from_node(path=Test_Zookeeper_Path) == _new_value, \
